Split server and routes/config
Turn HtmlGeneratorPlugin into a module
Turn server into dev-server and simplify it
