+++ webpack.config.js
@@ -1,39 +1,10 @@
 var autoprefixer = require('autoprefixer');
 var CopyWebpackPlugin = require('copy-webpack-plugin');
+var MustacheRendererPlugin = require('./mustache-renderer-webpack-plugin');
 var path = require('path');
 var webpack = require('webpack');
 
-var routes = require('./server/routes.json');
-
-
-function HtmlGeneratorPlugin (options) {
-    this.options = options;
-    return this;
-}
-
-HtmlGeneratorPlugin.prototype.apply = function (compiler) {
-    var render = this.options.render;
-    var routes = this.options.routes;
-
-    compiler.plugin('emit', function (compilation, callback) {
-        var outputRoutes = {};
-        routes.forEach(function (route) {
-            var filename = route.view + '.html';
-            var content = render(route);
-            outputRoutes[route.pattern] = filename;
-            compilation.assets[filename] = {
-                source: function () {return content;},
-                size: function () {return content.length;}
-            };
-        });
-        var routeJson = JSON.stringify(outputRoutes);
-        compilation.assets['routes.json'] = {
-            source: function () {return routeJson;},
-            size: function () {return routeJson.length;}
-        };
-        callback();
-    });
-};
+var routes = require('./src/routes.json');
 
 // Prepare all entry points
 var entry = {
@@ -86,9 +57,10 @@ module.exports = {
         fs: 'empty'
     },
     plugins: [
-        new HtmlGeneratorPlugin({
-            render: require('./server/render.js'),
-            routes: routes
+        new MustacheRendererPlugin({
+            templatePath: path.resolve(__dirname, './src/template.html'),
+            routes: routes,
+            config: require('./src/template-config.js')
         }),
         new CopyWebpackPlugin([
             {from: 'static'},
